KratosSwimmingDEM.py is just a launcher. The algorithm contains everything
