+++ applications/swimming_DEM_application/python_scripts/swimming_DEM_algorithm.py
@@ -1,34 +1,71 @@
 from __future__ import print_function, absolute_import, division #makes KratosMultiphysics backward compatible with python 2.6 and 2.7
 
-#TODO: set self.SolverSettings only once! 
-#TODO: the same for self.fluid_model_part
 #TODO: test DEM bounding box
 
 import os
 import sys
+import math
 from KratosMultiphysics import *
 from KratosMultiphysics.DEMApplication import *
 from KratosMultiphysics.SwimmingDEMApplication import *
 import dem_main_script_ready_for_coupling as DEM_algorithm
+import CFD_DEM_coupling
 import swimming_DEM_procedures as SDP
+import swimming_DEM_gid_output
+import embedded
 import variables_management as vars_man
 import time as timer
 
+try:
+    import define_output  # MA: some GUI write this file, some others not!
+except ImportError:
+    pass
+
+# Import MPI modules if needed. This way to do this is only valid when using OpenMPI. For other implementations of MPI it will not work.
+if "OMPI_COMM_WORLD_SIZE" in os.environ:
+    # Kratos MPI
+    from KratosMultiphysics.MetisApplication import *
+    from KratosMultiphysics.MPISearchApplication import *
+    from KratosMultiphysics.mpi import *
+
+    # DEM Application MPI
+    import DEM_procedures_mpi as DEM_procedures
+    import DEM_material_test_script_mpi as DEM_material_test_script
+    print("Running under MPI...........")
+else:
+    # DEM Application
+    import DEM_procedures
+    import DEM_material_test_script
+    print("Running under OpenMP........")
 
 sys.path.insert(0,'')
-import DEM_explicit_solver_var as DEM_parameters
 BaseAlgorithm = DEM_algorithm.Solution
 
+
 class Algorithm(BaseAlgorithm):
-    def __init__(self, pp):
+    def __init__(self, varying_parameters = dict()):
         self.StartTimer()
+        self.main_path = os.getcwd()
+        
+        import ProjectParameters as pp        
         self.pp = pp
+        self.pp.main_path = os.getcwd()
+        
+        import DEM_explicit_solver_var as DEM_parameters
+        self.pp.CFD_DEM = DEM_parameters
+                                
+        self.SetCustomBetaParameters(varying_parameters)
         self.SetBetaParameters()
         self.SetDoSolveDEMVariable()
         # creating a basset_force tool to perform the operations associated with the calculation of this force along the path of each particle
         self.GetBassetForceTools()
         BaseAlgorithm.__init__(self)
         self.CreateParts()
+        
+    def Run(self):
+        self.Initialize()
+        self.RunMainTemporalLoop()
+        self.Finalize()
 
     def CreateParts(self):
         # Order must be respected here
@@ -55,7 +92,6 @@ class Algorithm(BaseAlgorithm):
         ##############################################################################
 
         #G
-        self.pp.CFD_DEM = DEM_parameters
         self.pp.CFD_DEM.fluid_already_calculated = 0
         self.pp.CFD_DEM.recovery_echo_level = 1
         self.pp.CFD_DEM.gradient_calculation_type = 1
@@ -118,7 +154,7 @@ class Algorithm(BaseAlgorithm):
     def SetDoSolveDEMVariable(self):
         self.pp.do_solve_dem = not self.pp.CFD_DEM.flow_in_porous_DEM_medium_option
 
-    def SetCustomBetaParamters(self, dictionary):
+    def SetCustomBetaParameters(self, dictionary):
         if len(dictionary) == 0:
             return
         else: # assign the specified values to the specified variables
@@ -146,7 +182,19 @@ class Algorithm(BaseAlgorithm):
         BaseAlgorithm.ReadModelParts(self, starting_node_Id, starting_elem_Id, starting_cond_Id)
 
     def Initialize(self):
-                              
+        
+        print("\nInitializing Problem...")
+        sys.stdout.flush()
+                
+        run_code = self.GetRunCode() ##MA: NOT USED?
+
+        # Moving to the recently created folder
+        os.chdir(self.main_path)
+        [self.post_path, data_and_results, self.graphs_path, MPI_results] = self.procedures.CreateDirectories(str(self.main_path), str(self.pp.CFD_DEM.problem_name))
+        SDP.CopyInputFilesIntoFolder(self.main_path, self.post_path)
+
+        self.mixed_model_part = self.all_model_parts.Get('MixedPart')
+
         self.vars_man = vars_man
         self.vars_man.ConstructListsOfVariables(self.pp)
 
@@ -154,6 +202,195 @@ class Algorithm(BaseAlgorithm):
         self.DispersePhaseInitialize()
                                 
         self.vars_man.AddingExtraProcessInfoVariables(self.pp, self.fluid_model_part, self.spheres_model_part)
+        
+        self.SetCutsOutput()        
+        
+        self.swimming_DEM_gid_io = swimming_DEM_gid_output.SwimmingDEMGiDOutput(self.pp.problem_name,
+                                                                           self.pp.VolumeOutput,
+                                                                           self.pp.GiDPostMode,
+                                                                           self.pp.GiDMultiFileFlag,
+                                                                           self.pp.GiDWriteMeshFlag,
+                                                                           self.pp.GiDWriteConditionsFlag)
+
+        self.swimming_DEM_gid_io.initialize_swimming_DEM_results(self.spheres_model_part, self.cluster_model_part, self.rigid_face_model_part, self.mixed_model_part)
+
+        self.SetDragOutput()
+          
+        self.SetPointGraphPrinter()
+        
+        self.TransferGravityFromDisperseToFluid()               
+
+        # coarse-graining: applying changes to the physical properties of the model to adjust for
+        # the similarity transformation if required (fluid effects only).
+        SDP.ApplySimilarityTransformations(self.fluid_model_part, self.pp.CFD_DEM.similarity_transformation_type , self.pp.CFD_DEM.model_over_real_diameter_factor )
+
+        # creating a Post Utils object that executes several post-related tasks
+        self.post_utils = SDP.PostUtils(self.swimming_DEM_gid_io,
+                                                       self.pp,
+                                                       self.fluid_model_part,
+                                                       self.spheres_model_part,
+                                                       self.cluster_model_part,
+                                                       self.rigid_face_model_part,
+                                                       self.mixed_model_part)
+
+        # creating an IOTools object to perform other printing tasks
+        self.io_tools = SDP.IOTools(self.pp)
+
+        # creating a projection module for the fluid-DEM coupling
+        h_min = 0.01
+        n_balls = 1
+        fluid_volume = 10
+        self.pp.CFD_DEM.n_particles_in_depth = int(math.sqrt(n_balls / fluid_volume)) # only relevant in 2D problems
+        # creating a physical calculations module to analyse the DEM model_part
+        dem_physics_calculator = SphericElementGlobalPhysicsCalculator(self.spheres_model_part)
+
+        if self.pp.CFD_DEM.coupling_level_type:
+
+            if self.pp.CFD_DEM.meso_scale_length <= 0.0 and self.spheres_model_part.NumberOfElements(0) > 0:
+                biggest_size = 2 * dem_physics_calculator.CalculateMaxNodalVariable(self.spheres_model_part, RADIUS)
+                self.pp.CFD_DEM.meso_scale_length  = 20 * biggest_size
+
+            elif self.spheres_model_part.NumberOfElements(0) == 0:
+                self.pp.CFD_DEM.meso_scale_length  = 1.0
+
+            field_utility = self.GetFieldUtility()
+
+            self.projection_module = CFD_DEM_coupling.ProjectionModule(self.fluid_model_part, self.spheres_model_part, self.rigid_face_model_part, self.pp.domain_size, self.pp, field_utility)
+            self.projection_module.UpdateDatabase(h_min)
+
+        # creating a custom functions calculator for the implementation of additional custom functions
+        self.custom_functions_tool = SDP.FunctionsCalculator(self.pp)
+
+        # creating a derivative recovery tool to calculate the necessary derivatives from the fluid solution (gradient, laplacian, material acceleration...)
+        self.derivative_recovery_tool = DerivativeRecoveryTool3D(self.fluid_model_part)
+
+        # creating a stationarity assessment tool
+        self.stationarity_tool = SDP.StationarityAssessmentTool(self.pp.CFD_DEM.max_pressure_variation_rate_tol , self.custom_functions_tool)
+
+        # creating a debug tool
+        self.dem_volume_tool = SDP.ProjectionDebugUtils(self.pp.CFD_DEM.fluid_domain_volume, self.fluid_model_part, self.spheres_model_part, self.custom_functions_tool)
+
+        # creating a distance calculation process for the embedded technology
+        # (used to calculate elemental distances defining the structure embedded in the fluid mesh)
+        if self.pp.CFD_DEM.embedded_option:
+            self.calculate_distance_process = CalculateSignedDistanceTo3DSkinProcess(self.rigid_face_model_part, self.fluid_model_part)
+            self.calculate_distance_process.Execute()
+
+        self.KRATOSprint("Initialization Complete" + "\n")
+        sys.stdout.flush()
+
+        self.step           = 0
+        self.time           = self.pp.Start_time
+        self.Dt             = self.pp.Dt
+        self.out            = self.Dt
+        Nsteps         = self.pp.nsteps
+        self.final_time     = self.pp.CFD_DEM.FinalTime
+        self.output_time    = self.pp.CFD_DEM.OutputTimeStep
+
+        self.report.Prepare(self.timer, self.pp.CFD_DEM.ControlTime)
+
+        #first_print = True; index_5 = 1; index_10 = 1; index_50 = 1; control = 0.0
+
+        if (self.pp.CFD_DEM.ModelDataInfo == "ON"):
+            os.chdir(data_and_results)
+            if (self.pp.CFD_DEM.ContactMeshOption == "ON"):
+                (coordination_number) = self.procedures.ModelData(self.spheres_model_part, self.solver) # Calculates the mean number of neighbours the mean radius, etc..
+                self.KRATOSprint ("Coordination Number: " + str(coordination_number) + "\n")
+                os.chdir(self.main_path)
+            else:
+                self.KRATOSprint("Activate Contact Mesh for ModelData information")
+
+        if self.pp.CFD_DEM.flow_in_porous_medium_option:
+            fluid_frac_util = SDP.FluidFractionFieldUtility(self.fluid_model_part, self.pp.CFD_DEM.min_fluid_fraction )
+
+            for field in self.pp.fluid_fraction_fields:
+                fluid_frac_util.AppendLinearField(field)
+
+            fluid_frac_util.AddFluidFractionField()
+
+        if self.pp.CFD_DEM.flow_in_porous_DEM_medium_option:
+            SDP.FixModelPart(self.spheres_model_part)
+
+        # choosing the directory in which we want to work (print to)
+
+        os.chdir(self.post_path)
+
+        
+
+        ######################################################################################################################################
+
+        #                      I N I T I A L I Z I N G    T I M E    L O O P     ...   ( M I X E D    F L U I D / D E M    B L O C K )
+
+        ######################################################################################################################################
+
+        # setting up loop counters: Counter(steps_per_tick_step, initial_step, active_or_inactive_boolean, dead_or_not)
+        self.fluid_solve_counter          = self.GetFluidSolveCounter()
+        self.embedded_counter             = self.GetEmbeddedCounter()
+        self.DEM_to_fluid_counter         = self.GetBackwardCouplingCounter()
+        self.derivative_recovery_counter  = self.GetRecoveryCounter()
+        self.stationarity_counter         = self.GetStationarityCounter()
+        self.print_counter                = self.GetPrintCounter()
+        self.debug_info_counter           = self.GetDebugInfo()
+        self.particles_results_counter    = self.GetParticlesResultsCounter()
+        self.quadrature_counter           = self.HistoryForceQuadratureCounter()
+        self.mat_deriv_averager           = SDP.Averager(1, 3)
+        self.laplacian_averager           = SDP.Averager(1, 3)
+
+        ##############################################################################
+        #                                                                            #
+        #    MAIN LOOP                                                               #
+        #                                                                            #
+        ##############################################################################
+
+        self.DEM_step     = 0      # necessary to get a good random insertion of particles   # relevant to the stationarity assessment tool
+        self.time_dem     = 0.0
+        self.Dt_DEM       = self.spheres_model_part.ProcessInfo.GetValue(DELTA_TIME)
+        self.rigid_face_model_part.ProcessInfo[DELTA_TIME] = self.Dt_DEM
+        self.cluster_model_part.ProcessInfo[DELTA_TIME] = self.Dt_DEM
+        self.stationarity = False
+
+        self.report.total_steps_expected = int(self.pp.CFD_DEM.FinalTime / self.Dt_DEM)
+
+        self.KRATOSprint(self.report.BeginReport(self.timer))
+
+        # creating a Post Utils object that executes several post-related tasks
+        self.post_utils_DEM = DEM_procedures.PostUtils(self.pp.CFD_DEM, self.spheres_model_part)
+
+        SDP.InitializeVariablesWithNonZeroValues(self.fluid_model_part, self.spheres_model_part, self.pp) # otherwise variables are set to 0 by default
+
+        self.SetUpResultsDatabase()
+
+        # ANALYTICS BEGIN
+        self.pp.CFD_DEM.perform_analytics_option = False
+
+        if self.pp.CFD_DEM.perform_analytics_option:
+            import analytics
+            variables_to_measure = [PRESSURE]
+            steps_between_measurements = 100
+            gauge = analytics.Gauge(self.fluid_model_part, self.Dt, self.final_time, variables_to_measure, steps_between_measurements)
+            point_coors = [0.0, 0.0, 0.01]
+            target_node = SDP.FindClosestNode(self.fluid_model_part, point_coors)
+            target_id = target_node.Id
+            print(target_node.X, target_node.Y, target_node.Z)
+            print(target_id)
+            def condition(node):
+                return node.Id == target_id
+
+            gauge.ConstructArrayOfNodes(condition)
+            print(gauge.variables)
+            print_analytics_counter = SDP.Counter( 5 * steps_between_measurements, 1, 1)
+        # ANALYTICS END
+
+        import derivative_recovery.derivative_recovery_strategy as derivative_recoverer
+        self.recovery = derivative_recoverer.DerivativeRecoveryStrategy(self.pp, self.fluid_model_part, self.derivative_recovery_tool, self.custom_functions_tool)
+
+        self.FillHistoryForcePrecalculatedVectors()
+
+        self.PerformZeroStepInitializations()
+
+        self.post_utils.Writeresults(self.time)
+                              
+        
 
     def FluidInitialize(self):
         
@@ -202,14 +439,163 @@ class Algorithm(BaseAlgorithm):
         self.SetFluidBufferSizeAndAddDofsByFluidSolver()
         self.SetFluidBufferSizeAndAddDofsBySwimmingDEMAlgorithm()  
 
-    def SetFluidBufferSizeAndAddDofsByFluidSolver(self):
-        #spheres_model_part = self.all_model_parts.Get('SpheresPart')        
-        self.SolverSettings = self.pp.FluidSolverConfiguration
+    def SetFluidBufferSizeAndAddDofsByFluidSolver(self):        
         self.fluid_model_part.SetBufferSize(3)
         self.solver_module.AddDofs(self.fluid_model_part, self.SolverSettings)
         
     def SetFluidBufferSizeAndAddDofsBySwimmingDEMAlgorithm(self):
         SDP.AddExtraDofs(self.pp, self.fluid_model_part, self.spheres_model_part, self.cluster_model_part, self.DEM_inlet_model_part)
+        
+    def RunMainTemporalLoop(self):
+
+        while (self.time <= self.final_time):
+
+            self.time = self.time + self.Dt
+            self.step += 1
+            self.fluid_model_part.CloneTimeStep(self.time)
+            self.TellTime(self.time)
+
+
+            if self.pp.CFD_DEM.coupling_scheme_type  == "UpdatedDEM":
+                time_final_DEM_substepping = self.time + self.Dt
+
+            else:
+                time_final_DEM_substepping = self.time
+
+            # calculating elemental distances defining the structure embedded in the fluid mesh
+            if self.pp.CFD_DEM.embedded_option:
+                self.calculate_distance_process.Execute()
+
+            if self.embedded_counter.Tick():
+                embedded.ApplyEmbeddedBCsToFluid(self.fluid_model_part)
+                embedded.ApplyEmbeddedBCsToBalls(self.spheres_model_part, self.pp.CFD_DEM)
+
+            # solving the fluid part
+
+            if self.step >= 3 and not self.stationarity:
+                print("Solving Fluid... (", self.fluid_model_part.NumberOfElements(0), "elements )")
+                sys.stdout.flush()
+
+                if self.fluid_solve_counter.Tick():
+                    self.FluidSolve(self.time)
+
+            # assessing stationarity
+
+                if self.stationarity_counter.Tick():
+                    print("Assessing Stationarity...")
+                    sys.stdout.flush()
+                    self.stationarity = self.stationarity_tool.Assess(self.fluid_model_part)
+                    sys.stdout.flush()
+
+            # printing if required
+
+            if self.particles_results_counter.Tick():
+                # eliminating remote balls
+
+                #if self.pp.dem.BoundingBoxOption == "ON":
+                #    self.creator_destructor.DestroyParticlesOutsideBoundingBox(self.spheres_model_part)
+
+                self.io_tools.PrintParticlesResults(self.pp.variables_to_print_in_file, self.time, self.spheres_model_part)
+                #self.graph_printer.PrintGraphs(self.time) #MA: commented out because the constructor was already commented out
+                self.PrintDrag(self.drag_list, self.drag_file_output_list, self.fluid_model_part, self.time)
+
+            if self.output_time <= self.out and self.pp.CFD_DEM.coupling_scheme_type == "UpdatedDEM":
+
+                if self.pp.CFD_DEM.coupling_level_type > 0:
+                    self.projection_module.ComputePostProcessResults(self.spheres_model_part.ProcessInfo)
+
+                self.post_utils.Writeresults(self.time)
+                self.out = 0
+
+            # solving the DEM part
+            self.derivative_recovery_counter.Switch(self.time > self.pp.CFD_DEM.interaction_start_time)
+
+            if self.derivative_recovery_counter.Tick():
+                self.recovery.Recover()
+
+            print("Solving DEM... (", self.spheres_model_part.NumberOfElements(0), "elements )")
+            sys.stdout.flush()
+            first_dem_iter = True
+
+            for self.time_dem in self.yield_DEM_time(self.time_dem, time_final_DEM_substepping, self.Dt_DEM):
+                self.DEM_step += 1   # this variable is necessary to get a good random insertion of particles
+                self.spheres_model_part.ProcessInfo[TIME_STEPS]    = self.DEM_step
+                self.rigid_face_model_part.ProcessInfo[TIME_STEPS] = self.DEM_step
+                self.cluster_model_part.ProcessInfo[TIME_STEPS]    = self.DEM_step
+
+                self.PerformInitialDEMStepOperations(self.time_dem)
+
+                if self.time >= self.pp.CFD_DEM.interaction_start_time and self.pp.CFD_DEM.coupling_level_type and (self.pp.CFD_DEM.project_at_every_substep_option or first_dem_iter):
+
+                    if self.pp.CFD_DEM.coupling_scheme_type == "UpdatedDEM":
+                        self.ApplyForwardCoupling()
+
+                    else:
+                        self.ApplyForwardCoupling((time_final_DEM_substepping - self.time_dem) / self.Dt)
+
+                        if self.pp.CFD_DEM.IntegrationScheme in {'Hybrid_Bashforth', 'TerminalVelocityScheme'}:
+                            self.solver.Solve() # only advance in space
+                            self.ApplyForwardCouplingOfVelocityOnly(self.time_dem)
+                        else:
+                            if self.pp.CFD_DEM.basset_force_type > 0:
+                                node.SetSolutionStepValue(SLIP_VELOCITY_X, vx)
+                                node.SetSolutionStepValue(SLIP_VELOCITY_Y, vy)
+
+                        if self.quadrature_counter.Tick():
+                            self.AppendValuesForTheHistoryForce()
+
+                # performing the time integration of the DEM part
+
+                self.spheres_model_part.ProcessInfo[TIME]    = self.time_dem
+                self.rigid_face_model_part.ProcessInfo[TIME] = self.time_dem
+                self.cluster_model_part.ProcessInfo[TIME]    = self.time_dem
+
+                if self.pp.do_solve_dem:
+                    self.DEMSolve(self.time_dem)
+
+                self.DEMFEMProcedures.MoveAllMeshes(self.all_model_parts, self.time_dem, self.Dt_DEM)
+
+                #### TIME CONTROL ##################################
+
+                # adding DEM elements by the inlet:
+                if self.pp.CFD_DEM.dem_inlet_option:
+                    self.DEM_inlet.CreateElementsFromInletMesh(self.spheres_model_part, self.cluster_model_part, self.creator_destructor)  # After solving, to make sure that neighbours are already set.
+
+                if self.output_time <= self.out and self.pp.CFD_DEM.coupling_scheme_type == "UpdatedFluid":
+
+                    if self.pp.CFD_DEM.coupling_level_type:
+                        self.projection_module.ComputePostProcessResults(self.spheres_model_part.ProcessInfo)
+
+                    self.post_utils.Writeresults(self.time_dem)
+                    self.out = 0
+
+                self.out = self.out + self.Dt_DEM
+                first_dem_iter = False
+
+                # applying DEM-to-fluid coupling
+
+                if self.DEM_to_fluid_counter.Tick() and self.time >= self.pp.CFD_DEM.interaction_start_time:
+                    self.projection_module.ProjectFromParticles()
+
+            #### PRINTING GRAPHS ####
+            os.chdir(self.graphs_path)
+            # measuring mean velocities in a certain control volume (the 'velocity trap')
+            if self.pp.CFD_DEM.VelocityTrapOption:
+                self.post_utils_DEM.ComputeMeanVelocitiesinTrap("Average_Velocity.txt", self.time)
+
+            os.chdir(self.post_path)
+
+            # coupling checks (debugging)
+            if self.debug_info_counter.Tick():
+                self.dem_volume_tool.UpdateDataAndPrint(self.pp.CFD_DEM.fluid_domain_volume)
+
+            # printing if required
+
+            if self.particles_results_counter.Tick():
+                self.io_tools.PrintParticlesResults(self.pp.variables_to_print_in_file, self.time, self.spheres_model_part)
+                #self.graph_printer.PrintGraphs(self.time) #MA: commented out because the constructor was already commented out
+                self.PrintDrag(self.drag_list, self.drag_file_output_list, self.fluid_model_part, self.time)
+                
         
     def DEMSolve(self, time = 'None'): # time is passed in case it is needed
         self.solver.Solve()
@@ -224,7 +610,7 @@ class Algorithm(BaseAlgorithm):
         pass
 
     def SetInlet(self):
-        if DEM_parameters.dem_inlet_option:
+        if self.pp.CFD_DEM.dem_inlet_option:
             #Constructing the inlet and initializing it (must be done AFTER the self.spheres_model_part Initialize)
             # Note that right now only inlets of a single type are possible. This should be generalized.
             if self.pp.type_of_inlet == 'VelocityImposed':
@@ -331,7 +717,6 @@ class Algorithm(BaseAlgorithm):
         self.basset_force_tool = SDP.BassetForceTools()
 
     def ActivateTurbulenceModel(self):
-        self.fluid_model_part = self.all_model_parts.Get('FluidPart')
 
         if self.pp.FluidSolverConfiguration.TurbulenceModel == "Spalart-Allmaras":
             # apply the initial turbulent viscosity on all of the nodes
@@ -369,6 +754,96 @@ class Algorithm(BaseAlgorithm):
 
     def PerformFinalOperations(self, time = None):
         pass
+    
+    def Finalize(self):
+        
+        self.swimming_DEM_gid_io.finalize_results()
+
+        self.PerformFinalOperations(self.time_dem)
+
+        for i in self.drag_file_output_list:
+            i.close()
+
+        self.TellFinalSummary(self.step, self.time, self.DEM_step)
+
+        return self.GetReturnValue()
+    
+    
+    def SetDragOutput(self):
+        # define the drag computation list
+        self.drag_list = define_output.DefineDragList()
+        self.drag_file_output_list = []
+        for it in self.drag_list:
+            f = open(it[1], 'w')
+            self.drag_file_output_list.append(f)
+            tmp = "#Drag for group " + it[1] + "\n"
+            f.write(tmp)
+            tmp = "time RX RY RZ"
+            f.write(tmp)
+            f.flush()
+
+        print(self.drag_file_output_list)  
+        sys.stdout.flush()
+    
+    def SetPointGraphPrinter(self):
+        pass
+         # preparing output of point graphs
+        #import point_graph_printer
+
+        #output_nodes_list = define_output.DefineOutputPoints()
+        #self.graph_printer = point_graph_printer.PrintGraphPrinter(
+            #output_nodes_list,
+            #fluid_model_part,
+            #variables_dictionary,
+            #domain_size)
+    
+    def SetCutsOutput(self):            
+        if not self.pp.VolumeOutput:
+            cut_list = define_output.DefineCutPlanes()
+            self.swimming_DEM_gid_io.define_cuts(self.fluid_model_part, cut_list)
+            
+    def TransferGravityFromDisperseToFluid(self):
+        # setting fluid's body force to the same as DEM's
+        if self.pp.CFD_DEM.body_force_on_fluid_option:
+
+            for node in self.fluid_model_part.Nodes:
+                node.SetSolutionStepValue(BODY_FORCE_X, 0, self.pp.CFD_DEM.GravityX)
+                node.SetSolutionStepValue(BODY_FORCE_Y, 0, self.pp.CFD_DEM.GravityY)
+                node.SetSolutionStepValue(BODY_FORCE_Z, 0, self.pp.CFD_DEM.GravityZ)
+    
+    def yield_DEM_time(self, current_time, current_time_plus_increment, delta_time):
+            current_time += delta_time
+
+            tolerance = 0.0001
+            while current_time < (current_time_plus_increment - tolerance * delta_time):
+                yield current_time
+                current_time += delta_time
+
+            current_time = current_time_plus_increment
+            yield current_time
+            
+    def PrintDrag(self, drag_list, drag_file_output_list, fluid_model_part, time):
+        i = 0
+        for it in drag_list:
+            print(it[0])
+            nodes = self.fluid_model_part.GetNodes(it[0])
+            drag = Vector(3)
+            drag[0] = 0.0
+            drag[1] = 0.0
+            drag[2] = 0.0
+            for node in nodes:
+                reaction = node.GetSolutionStepValue(REACTION, 0)
+                drag[0] += reaction[0]
+                drag[1] += reaction[1]
+                drag[2] += reaction[2]
+
+            output = str(time) + " " + str(drag[0]) + " " + str(drag[1]) + " " + str(drag[2]) + "\n"
+            # print self.drag_file_output_list[i]
+            # print output
+            self.drag_file_output_list[i].write(output)
+            self.drag_file_output_list[i].flush()
+            i = i + 1
+    
     
     def GetReturnValue(self):
         return 0.0
